pyproject.toml
[project]
name = "datapipeline-z-image"
version = "0.1.0"
description = "Data pipeline for Z-Image: Efficient Image Generation Foundation Model"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.5.0",
    "ffmpeg-python>=0.2.0",
    "imagehash>=4.3.2",
    "numpy>=1.24.0",
    "pillow>=12.1.0",
    "pyarrow>=14.0.0",
    "pyiceberg>=0.7.0",
    "pyyaml>=6.0.0",
    "ray[data,default]>=2.53.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "open-clip-torch>=3.2.0",
]

[project.optional-dependencies]
enhanced = [
    "langdetect>=1.0.9",    # For better language detection
    "pytesseract>=0.3.10",  # For OCR text detection
    "opencv-python>=4.8.0", # For advanced image processing
]
ml = [
    "torch>=2.0.0",            # PyTorch for deep learning models
    "torchvision>=0.15.0",     # Torchvision for image processing
    "open-clip-torch>=2.20.0", # OpenCLIP for image embeddings
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["framework", "operators", "loaders", "writers", "models"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
